+++ packages/effects/hooks/src/use-design-tokens.ts
@@ -1,9 +1,7 @@
 import { reactive, watch } from 'vue';
 
 import { preferences } from '@vben/preferences';
-import { updateCSSVariables } from '@vben/utils';
-
-import { hlsStringToRGB } from './convert';
+import { hlsStringToRGBSting, updateCSSVariables } from '@vben/utils';
 
 /**
  * 用于适配各个框架的设计系统
@@ -104,10 +102,7 @@ export function useNaiveDesignTokens() {
 
   const getCssVariableValue = (variable: string, isColor: boolean = true) => {
     const value = rootStyles.getPropertyValue(variable);
-    const rgbColor = hlsStringToRGB(`hsl(${value})`);
-    const { a, b = 0, g = 0, r = 0 } = { a: 1, ...rgbColor };
-
-    return isColor ? `rgba(${r}, ${g}, ${b}, ${a})` : value;
+    return isColor ? hlsStringToRGBSting(value) : value;
   };
 
   watch(
